+++ test/src/animation_id_test.dart
@@ -1,3 +1,4 @@
+import 'package:flutter_map/plugin_api.dart';
 import 'package:flutter_map_animations/src/animation_id.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:latlong2/latlong.dart';
@@ -5,11 +6,12 @@ import 'package:latlong2/latlong.dart';
 void main() {
   group('AnimationId', () {
     const validId = 'started#51.0,0.0,1.0';
+    const validIdWithCustomId = 'started#51.0,0.0,1.0#aCustomId';
     const invalidId = 'invalid#51.0,0.0,1.0';
 
     group('id', () {
       test('should return a formatted string', () {
-        final animationId = AnimationId(
+        const animationId = AnimationId(
           destLocation: LatLng(51, 0),
           destZoom: 1,
         );
@@ -23,6 +25,19 @@ void main() {
         final animationId = AnimationId.parse(validId);
 
         expect(animationId, isA<AnimationId>());
+      });
+
+      test('should return a customId', () {
+        final animationId = AnimationId.parse(validIdWithCustomId);
+
+        expect(
+          animationId,
+          isA<AnimationId>().having(
+            (animationId) => animationId.customId,
+            'customId',
+            'aCustomId',
+          ),
+        );
       });
 
       test('should throw an error if the moveId is invalid', () {
@@ -39,6 +54,36 @@ void main() {
 
       test('should return null if the moveId is invalid', () {
         final animationId = AnimationId.tryParse(invalidId);
+
+        expect(animationId, isNull);
+      });
+    });
+
+    group('fromMapEvent', () {
+      test('should return an AnimationId', () {
+        final animationId = AnimationId.fromMapEvent(
+          const MapEventMove(
+            center: LatLng(1, 2),
+            zoom: 6,
+            targetCenter: LatLng(2, 4),
+            targetZoom: 5,
+            source: MapEventSource.custom,
+            id: validId,
+          ),
+        );
+
+        expect(animationId, isNotNull);
+      });
+
+      test('should return null if the MapEvent is not a MapEventMove', () {
+        final animationId = AnimationId.fromMapEvent(
+          const MapEventTap(
+            center: LatLng(1, 2),
+            zoom: 6,
+            source: MapEventSource.custom,
+            tapPosition: LatLng(3, 4),
+          ),
+        );
 
         expect(animationId, isNull);
       });